Add gradient estimation class and test case.
This class estimates gradients in discrete space by mapping
tokens to embeddings and estimating gradients in embedding
space. Replacing one token with another is equivalent to
perturbing the embedding of a token with a displacement vector.
Scaling the displacement vector by the difference in loss
yields a vector that points in the direction of increased loss.

Change-Id: Id43a6ff8a32cf7c70922989b5d355a44aded0673