Added importance scoring and token dropping helper functions.
In order to decide which tokens to change we need to know which
tokens affect the output of our model the most. The scorer
function works by iterating over each token in the sentences and
either masking them or deleting them and then computing the
difference in model output, given by a forward function.

Change-Id: Ie394411fbcdcca720097af298cf9c530e3151564