Adds Universal Sentence Encoder for semantic loss.
Previous semantic similarity constraints weren't powerful enough
to prevent the attack from producing garbage tokens. Other papers
have used the Universal Sentence Encoder as their semantic
similarity constraint. We hope that the attack can't game it as
easily as it appears to be possible with cosine similarity and
euclidean distance.

Change-Id: I441e682560e99499e4081fda2252950d70fb5fda