Adds option to add displacement to embedding before discretization.
Previously we assumed that the mean displacement is a sort of
gradient and that multiplying it with the embeddings would yield
token scores, like multiplying the gradient with the embeddings in
hotflip does. If this does not hold, it could be useful to try
adding the displacement to the current embeddings and then find
the most similar embeddings by cosine or maximum inner product.

Change-Id: Ice549604436dbc96a5c14613c998ac22f62601e1