Prevents artificially high query counts.
This commit adds a check inside of estimation.py to see if the
token we will be replacing is from a sentence that has been
removed from the batch. This is recognizable by it being a padding
token. When it is, then we won't query the model with replacements
for that token, saving query counts. Additionally, we prevent
nan importance scores from appearing and affecting which tokens
are attacked.

Change-Id: Ib230037d95d2f5ca128e5bcb6996d00862f44731